Decouple the closure based glyph segmentation functionality from GlyphSegmentation.
- GlyphSegmentation contains common structures around representing a segmentation.
- The specific closure based implementation of glyph keyed segmentation is moved to it's own file and own util closure_glyph_segmenter/closure_glyph_keyed_segmenter_util.
