+++ ift/encoder/glyph_segmentation.h
@@ -24,9 +24,9 @@ typedef uint32_t glyph_id_t;
  * patches.
  *
  * A segmentation describes the groups of glyphs belong to each patch as well as
- * the conditions under which those patches should be loaded. This gaurantees
- * that the produced set of patches and conditions will satisfy the "glyph
- * closure requirement", which is:
+ * the conditions under which those patches should be loaded. A properly formed
+ * segmentation should have an associated set of patches and conditions which will
+ * satisfy the "glyph closure requirement", which is:
  *
  * The set of glyphs contained in patches loaded for a font subset definition (a
  * set of Unicode codepoints and a set of layout feature tags) through the patch
@@ -35,8 +35,9 @@ typedef uint32_t glyph_id_t;
  */
 class GlyphSegmentation {
  public:
-  // TODO(garretrieger): merge this with Encoder::Condition they are basically
-  // identical.
+  /*
+   * The conditions under which a group of glyphs should be laoded.
+   */
   class ActivationCondition {
    public:
     /*
@@ -125,6 +126,11 @@ class GlyphSegmentation {
     patch_id_t activated_;
   };
 
+  GlyphSegmentation(absl::btree_set<glyph_id_t> init_font_glyphs,
+                    absl::btree_set<glyph_id_t> unmapped_glyphs)
+      : init_font_glyphs_(init_font_glyphs),
+        unmapped_glyphs_(unmapped_glyphs) {}
+
   /*
    * Converts a list of activation conditions into a list of condition entries
    * which are used by the encoder to specify conditions.
@@ -134,21 +140,6 @@ class GlyphSegmentation {
       absl::Span<const ActivationCondition> conditions,
       const absl::flat_hash_map<segment_index_t,
                                 absl::flat_hash_set<hb_codepoint_t>>& segments);
-
-  /*
-   * Analyzes a set of codepoint segments using a subsetter closure and computes
-   * a GlyphSegmentation which will satisfy the "glyph closure requirement" for
-   * the provided font face.
-   *
-   * initial_segment is the set of codepoints that will be placed into the
-   * initial ift font.
-   */
-  // TODO(garretrieger): also support optional feature segments.
-  static absl::StatusOr<GlyphSegmentation> CodepointToGlyphSegments(
-      hb_face_t* face, absl::flat_hash_set<hb_codepoint_t> initial_segment,
-      std::vector<absl::flat_hash_set<hb_codepoint_t>> codepoint_segments,
-      uint32_t patch_size_min_bytes = 0,
-      uint32_t patch_size_max_bytes = UINT32_MAX);
 
   /*
    * Returns a human readable string representation of this segmentation and
@@ -201,7 +192,6 @@ class GlyphSegmentation {
     return init_font_glyphs_;
   };
 
- private:
   static absl::Status GroupsToSegmentation(
       const absl::btree_map<absl::btree_set<segment_index_t>,
                             absl::btree_set<glyph_id_t>>& and_glyph_groups,
@@ -212,10 +202,7 @@ class GlyphSegmentation {
 
   void CopySegments(const std::vector<common::hb_set_unique_ptr>& segments);
 
-  // TODO(garretrieger): the output conditions need to also capture the base
-  // codepoint segmentations since those
-  //                     form the base conditions which composite conditions are
-  //                     built up from.
+ private:
   absl::btree_set<glyph_id_t> init_font_glyphs_;
   absl::btree_set<glyph_id_t> unmapped_glyphs_;
   absl::btree_set<ActivationCondition> conditions_;
